refactor(runtime): move provider URL helpers to provider-urls.ts
`subscription-auth.ts` was holding `anthropicRootUrl` / `anthropicV1BaseUrl` /
`anthropicV1Url` / `googleV1BetaBaseUrl` / `googleApiUrl` / `stripTrailing` —
none of which is subscription auth. Move them to `provider-urls.ts` so the file
name matches its content; `subscription-auth.ts` keeps only the auth headers
and the Codex JWT decode.

Update the five importers to split the URL helpers from the auth headers, and
fix the missing trailing newline in `google-base-url.test.ts`. The local
`stripTrailing` in `test-connection.ts` and `model-fetcher.ts` is independent
and left as-is.
